V3R_Studio/ImageCaptures/requirements.txt
numpy==1.25.0
piexif==1.1.3
Pillow==10.0.0
